feat: refactor event handling and recently viewed functionality with modular scripts
