+++ event-detail.html
@@ -10,7 +10,7 @@
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
     />
     <link rel="stylesheet" href="css/main.css" />
-    <script src="js/event-detail-page.js" defer></script>
+    <script type="module" src="js/event-detail-page.js" defer></script>
   </head>
   <body>
     <!-- Desktop and tablet header -->
@@ -55,18 +55,19 @@
       </div>
     </header>
     <main>
-      <section class="event-detail-page"></section>
-      <div class="recently-viewed-slide">
-        <h3 class="home-title">Recently viewed</h3>
+      <section class="event-detail-page">
+        <div class="recently-viewed-slide">
+          <h3 class="home-title">Recently viewed</h3>
 
-        <button class="rv-prev">
-          <i class="bi bi-chevron-left"></i>
-        </button>
-        <button class="rv-next">
-          <i class="bi bi-chevron-right"></i>
-        </button>
-        <div id="recently-viewed"></div>
-      </div>
+          <button class="rv-prev">
+            <i class="bi bi-chevron-left"></i>
+          </button>
+          <button class="rv-next">
+            <i class="bi bi-chevron-right"></i>
+          </button>
+          <div id="recently-viewed"></div>
+        </div>
+      </section>
     </main>
     <footer>
       <img src="img/Logo2.png" alt="" />
