+++ home.html
@@ -10,7 +10,7 @@
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
     />
     <link rel="stylesheet" href="css/main.css" />
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
     <!-- or detailPage.js -->
   </head>
   <body>
